moved most Ruby code into lib/, left a smaller check script at the root
